refactor: Move 'throttle' functionality to only place where it is used
